integrations-and-supported-tools/sacred/scripts/requirements.txt
neptune-client==0.9.18
neptune-sacred==0.9.3
sacred==0.8.2
torch==1.9.0
torchvision==0.10.0

